Move ScanArgs to query.go, move valueOf from decode to ScanArgs, move markerIndex to prepare.go
