+++ internal/expr/query.go
@@ -85,3 +85,66 @@ func (pe *PreparedExpr) Query(args ...any) (ce *QueryExpr, err error) {
 
 	return &QueryExpr{outputs: pe.outputs, sql: pe.sql, args: qargs}, nil
 }
+
+// ScanArgs returns list of pointers to the struct fields that are listed in qe.outputs.
+// All the structs mentioned in the query must be in outputArgs.
+// All outputArgs must be structs.
+func (qe *QueryExpr) ScanArgs(columns []string, outputArgs []any) ([]any, error) {
+	outputVals := []reflect.Value{}
+	for _, outputArg := range outputArgs {
+		if outputArg == nil {
+			return nil, fmt.Errorf("need pointer to struct, got nil")
+		}
+		outputVal := reflect.ValueOf(outputArg)
+		if outputVal.Kind() != reflect.Pointer {
+			return nil, fmt.Errorf("need pointer to struct, got %s", outputVal.Kind())
+		}
+		if outputVal.IsNil() {
+			return nil, fmt.Errorf("got nil pointer")
+		}
+		outputVal = reflect.Indirect(outputVal)
+		if outputVal.Kind() != reflect.Struct {
+			return nil, fmt.Errorf("need pointer to struct, got pointer to %s", outputVal.Kind())
+		}
+		outputVals = append(outputVals, outputVal)
+	}
+
+	// Check that each outputVal is in the query.
+	var inQuery = make(map[reflect.Type]bool)
+	for _, field := range qe.outputs {
+		inQuery[field.structType] = true
+	}
+	var typeDest = make(map[reflect.Type]reflect.Value)
+	for _, outputVal := range outputVals {
+		if !inQuery[outputVal.Type()] {
+			return nil, fmt.Errorf("type %q does not appear as an output type in the query", outputVal.Type().Name())
+		}
+		typeDest[outputVal.Type()] = outputVal
+	}
+
+	var ptrs = []any{}
+	for _, column := range columns {
+		idx, ok := markerIndex(column)
+		if !ok {
+			// Columns not mentioned in output expressions are scanned into x.
+			var x any
+			ptrs = append(ptrs, &x)
+			continue
+		}
+		if idx >= len(qe.outputs) {
+			return nil, fmt.Errorf("internal error: sqlair column not in outputs (%d>=%d)", idx, len(qe.outputs))
+		}
+		field := qe.outputs[idx]
+		outputVal, ok := typeDest[field.structType]
+		if !ok {
+			return nil, fmt.Errorf("type %q found in query but not passed to decode", field.structType.Name())
+		}
+
+		val := outputVal.FieldByIndex(field.index)
+		if !val.CanSet() {
+			return nil, fmt.Errorf("internal error: cannot set field %s of struct %s", field.name, field.structType.Name())
+		}
+		ptrs = append(ptrs, val.Addr().Interface())
+	}
+	return ptrs, nil
+}
